refactor: extract Telegram routes to routes/integrations/telegram.ts
